+++ pipdownload/__init__.py
@@ -1,7 +1,7 @@
 import logging
 import sys
 
-__version__ = '0.0.3'
+__version__ = '0.0.4'
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)