+++ tests/commands/test_cli.py
@@ -25,10 +25,15 @@ def test_download_from_requirement_file(requirement_file_path):
     with TempDirectory(delete=True) as directory:
         runner = CliRunner()
         result = runner.invoke(pipdownload, ['-r', requirement_file_path, '-d', directory.path])
-        print(requirement_file_path)
-        print(result.output)
-        print(result)
         assert result.exit_code == 0
         files = os.listdir(directory.path)
         assert len(files) == 2
 
+
+def test_download_with_option_whl_suffixes():
+    with TempDirectory(delete=True) as directory:
+        runner = CliRunner()
+        result = runner.invoke(pipdownload, ['grpcio==1.21.1', '-d', directory.path])
+        assert result.exit_code == 0
+        files = os.listdir(directory.path)
+        assert len(files) == 14